+++ backend/index.js
@@ -11,20 +11,20 @@ app.use(cors({
 app.use(express.json()); //to convert request data to json
 
 // Credential Apis
-app.use("/api/student/auth", require("./routes/Student Api/studentCredential"));
-app.use("/api/faculty/auth", require("./routes/Faculty Api/facultyCredential"));
-app.use("/api/admin/auth", require("./routes/Admin Api/adminCredential"));
+app.use("/api/student/auth", require("./routes/Student Api/credential.route"));
+app.use("/api/faculty/auth", require("./routes/Faculty Api/credential.route"));
+app.use("/api/admin/auth", require("./routes/Admin Api/credential.route"));
 // Details Apis
-app.use("/api/student/details", require("./routes/Student Api/studentDetails"));
-app.use("/api/faculty/details", require("./routes/Faculty Api/facultyDetails"));
-app.use("/api/admin/details", require("./routes/Admin Api/adminDetails"));
+app.use("/api/student/details", require("./routes/Student Api/details.route"));
+app.use("/api/faculty/details", require("./routes/Faculty Api/details.route"));
+app.use("/api/admin/details", require("./routes/Admin Api/details.route"));
 // Other Apis
-app.use("/api/timetable", require("./routes/timetable"));
-app.use("/api/material", require("./routes/material"));
-app.use("/api/notice", require("./routes/notice"));
-app.use("/api/subject", require("./routes/subject"));
-app.use("/api/marks", require("./routes/marks"));
-app.use("/api/branch", require("./routes/branch"));
+app.use("/api/timetable", require("./routes/Other Api/timetable.route"));
+app.use("/api/material", require("./routes/Other Api/material.route"));
+app.use("/api/notice", require("./routes/Other Api/notice.route"));
+app.use("/api/subject", require("./routes/Other Api/subject.route"));
+app.use("/api/marks", require("./routes/Other Api/marks.route"));
+app.use("/api/branch", require("./routes/Other Api/branch.route"));
 
 app.listen(port, () => {
   console.log(`Server Listening On http://localhost:${port}`);
