+++ backend/routes/Other Api/material.route.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+const { getMaterial, addMaterial, updateMaterial, deleteMaterial } = require("../../controllers/Other/material.controller.js");
+
+router.post("/getMaterial", getMaterial);
+router.post("/addMaterial", addMaterial);
+router.put("/updateMaterial/:id", updateMaterial);
+router.delete("/deleteMaterial/:id", deleteMaterial);
+
+module.exports = router;
