+++ backend/routes/Other Api/notice.route.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const { getNotice, addNotice, updateNotice, deleteNotice } = require("../../controllers/Other/notice.controller");
+const router = express.Router();
+
+router.post("/getNotice", getNotice);
+router.post("/addNotice", addNotice);
+router.put("/updateNotice/:id", updateNotice);
+router.delete("/deleteNotice/:id", deleteNotice);
+
+module.exports = router;
