+++ backend/routes/Other Api/timetable.route.js
@@ -1,0 +1,11 @@
+require("dotenv").config();
+const express = require("express");
+const router = express.Router();
+const upload = require("../../middlewares/multer.middleware.js");
+const { getTimetable, addTimetable, deleteTimetable } = require("../../controllers/Other/timetable.controller.js");
+
+router.post("/getTimetable", getTimetable);
+router.post("/addTimetable", upload.single("file"), addTimetable);
+router.delete("/deleteTimetable/:id", deleteTimetable);
+
+module.exports = router;
