+++ backend/routes/Student Api/credential.route.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const router = express.Router();
+const { loginHandler, registerHandler, updateHandler, deleteHandler } = require("../../controllers/Student/credential.controller.js");
+
+router.post("/login", loginHandler);
+
+router.post("/register", registerHandler);
+
+router.put("/update/:id", updateHandler);
+
+router.delete("/delete/:id", deleteHandler);
+
+module.exports = router;
